// worklenz-frontend/src/app/administrator/reporting/interfaces.ts
export interface IRPTTaskDrawerData {
  taskId: string;
  projectId: string;
}